[xbuild] Add clang-cl cross-compilation support for Linux to Windows
Add CMake-based build system using clang-cl + lld-link + xwin toolchain
for cross-compiling from Linux to Windows (x64 and ARM64).

Build system:
- CMakeLists.txt with full target definitions
- CMakePresets.json for x64/ARM64 Release/Debug configurations
- Toolchain files for clang-cl cross-compilation

Source fixes for clang-cl compatibility:
- Fix C++11 access declarations (using CDialog::X syntax)
- Fix case-sensitive includes (uxtheme.lib, header files)
- Fix DWORD narrowing in switch case for negative constant
- Fix non-POD type through variadic function (CQuadWord.Value)
- Rename duplicate static symbol (IsUNCPathLocal)

Resource compiler (llvm-rc) compatibility:
- Fix case-sensitive RC file includes
- Add missing .rh placeholder files
- Normalize file naming (uppercase .RC -> lowercase .rc)

Successfully builds salamander.exe and 9 plugins on Linux.

+++ src/salamdr5.cpp
@@ -471,7 +471,7 @@ RETRY:
     {
         switch (lastError)
         {
-        case ERROR_USER_TERMINATED:
+        case (DWORD)ERROR_USER_TERMINATED:
             break;
 
         case ERROR_NOT_READY:
